backend/requirements.txt
django~=4.2.9
djangorestframework~=3.14.0
django-cors-headers